% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background_color.R
\name{background_color}
\alias{background_color}
\title{Modify ggplot theme background}
\usage{
background_color(color_bg = "white", color_border = "black")
}
\arguments{
\item{color_bg}{The desired background color.}

\item{color_border}{The desired border color.}
}
\description{
Easily change the background color for a ggplot. Changes both the panel
and plot backgrounds to the same desired \code{color_bg}. Can also change the
border color of the plot with \code{color_border}
}
\examples{
# Make example plot:
library(ggplot2)
data("midwest", package = "ggplot2")
plot <- ggplot(midwest, aes(x = area, y = poptotal)) +
    geom_point(aes(col = state, size = popdensity)) +
    geom_smooth(method = "loess", se = FALSE) +
    xlim(c(0, 0.1)) +
    ylim(c(0, 500000)) +
    labs(subtitle = "Area Vs Population",
        y = "Population",
        x = "Area",
        title = "Scatterplot",
        caption = "Source: midwest")
plot

# Change the background color to blue
plot + background_color('blue')

# Change the background color to white and border to blue
plot + background_color(color_bg = 'white', color_border = 'blue')
}
