% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concatCode.R
\name{concatCode}
\alias{concatCode}
\title{Prints the code to concatenate a set of strings}
\usage{
concatCode(strings, returns = FALSE)
}
\arguments{
\item{strings}{A set of strings.}

\item{returns}{Set to TRUE to print a new line in between each string}
}
\description{
This function prints the code to concatenate a set of strings.
}
\examples{
# Create a set of strings:
strings = c('one', 'two', 'one', 'three', 'two', 'four')

# Print code to create the concatenated set of all unique string values:
concatCode(unique(strings))

# Print code with new lines in between each string:
concatCode(unique(strings), returns=T)
}
