% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummyCode.R
\name{dummyCode}
\alias{dummyCode}
\title{Creates dummy coded variables in a data frame.}
\usage{
dummyCode(df, vars)
}
\arguments{
\item{vars}{The variables in the data frame for which you want to create new dummy coded variables}

\item{data}{A data frame.}
}
\description{
This function Creates dummy coded variables in a data frame.
}
\examples{
# Create an example data frame:
df = data.frame(
    animal   = c("dog", "goldfish", "bird", "dog", "goldfish"),
    numLegs  = c(4, 0, 2, 4, 0),
    lifeSpan = c(10, 10, 5, 10, 10))
df

# Create dummy coded variables for the variable "animal":
df.dummy = dummyCode(df, vars="animal"))
df.dummy
# Create dummy coded variables for the variables "animal" and "numLegs":
df.dummy = dummyCode(df, vars=c("animal", "numLegs"))
df.dummy

}
