% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMyPackageLibrary.R
\name{getMyPackageLibrary}
\alias{getMyPackageLibrary}
\title{Get a character list of all currently installed packages on your machine.}
\usage{
getMyPackageLibrary()
}
\description{
This function returns a character list of all currently installed packages
on your machine.
}
\examples{
packageList = getMyPackageLibrary()
}
