% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmultiplot.R
\name{ggmultiplot}
\alias{ggmultiplot}
\title{ggmultiplot}
\usage{
ggmultiplot(..., plotlist = NULL, file, cols = 1, layout = NULL)
}
\arguments{
\item{cols}{The number of columns in the layout}

\item{layout}{A matrix specifying the layout. If present, 'cols' is ignored.}
}
\description{
This function plots multiple ggplot objects into one plot frame.
}
\details{
A series of ggplot objects can be directly passed as ... or passed
If the layout is something like matrix(c(1,2,3,3), nrow=2, byrow=TRUE),
then plot 1 will go in the upper left, 2 will go in the upper right, and
3 will go all the way across the bottom.
}
