% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtornado.R
\name{ggTornado}
\alias{ggTornado}
\title{Create a "tornado" plot using ggplot2}
\usage{
ggTornado(
  data,
  baseline,
  varName = "var",
  levelName = "level",
  valName = "val",
  resultName = "result",
  xlab = "Result",
  ylab = "Parameter"
)
}
\arguments{
\item{data}{A data frame containing the results of a sensitivity analysis.}

\item{baseline}{The baseline value that the sensitivity analysis is compared to.}

\item{varName}{The data column identifying the variable being varied.}

\item{levelName}{The data column identifying the sensitivity case level (high or low, relative to the baseline).}

\item{valName}{The data column identifying the value of the variable being varied.}

\item{resultName}{The data column identifying the result of the output value at the varied variable value.}

\item{xlab}{The x-axis label (defaults to 'Result').}

\item{ylab}{The y-axis label (defaults to 'Parameter').}
}
\description{
This function creates a "tornado" plot using the ggplot2 package. These
are primarily used to display results of a sensitivity analysis.
}
\examples{

# Create an example data frame of a sensitivity analysis - columns:
# 'var'      = The name of the variable being varied.
# 'level'    = High or Low (relative to the baseline).
# 'varValue' = The value of the variable being varied.
# 'result'   = The result of the output value at the varied variable value.
data <- data.frame(
    var      = c('price', 'price', 'fuelEconomy', 'fuelEconomy',
                 'accelTime', 'accelTime'),
    level    = rep(c('high', 'low'), 3),
    varValue = c(10, 20, 25, 15, 10, 6),
    result   = c(0.95, 0.15, 0.90, 0.60, 0.85, 0.75))

# Define the baseline analysis result:
baseline = 0.8

# Make a tornado plot of the sensitivity analysis results:
library(tidyverse)
plot = ggTornado(data = data, baseline=baseline, varName='var',
                 levelName='level', valName='varValue',
                 resultName='result', xlab='Result',
                 ylab='Parameter')
plot
}
