% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installAllThePackages.R
\name{installAllThePackages}
\alias{installAllThePackages}
\title{Install loads of packages (>500) all at once.}
\usage{
installAllThePackages(preview = F)
}
\description{
This function installs a bunch of packages that I frequently use.
Useful after first installing or upgrading R.
}
\examples{
# View a list of all the packages:
installAllThePackages(preview=T)

# Install all the packages:
installAllThePackages()
}
