% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jColors.R
\name{jColors}
\alias{jColors}
\title{A list of nice color palettes.}
\usage{
jColors(palette = NULL, colors = NULL, preview = FALSE)
}
\arguments{
\item{palette}{A palette name. Options include: \code{"basic"}, \code{"extended"},
\code{"bright"}, \code{"stan"}, \code{"clean"}, \code{"set3"}, \code{"trove"}, \code{"redToGray"}.}

\item{colors}{Which colors you want from the palette?}

\item{preview}{Preview the colors in the palette?}
}
\description{
Returns a list of different color palettes. Each palette is a data.frame.
trove Palette found at http://colrd.com/palette/19308/
redToGray Palette found at http://colrd.com/palette/21894/
}
\examples{
# Preview every color palette
jColors(preview = TRUE)

# Preview a single color palette
jColors(palette = "basic", preview = TRUE)

# Preview a few colors from a single color palette
jColors(palette = "basic", colors = c("red", "blue"), preview = TRUE)

# Get the colors from a whole palette as a vector
jColors(palette = "basic")

# Get only select colors from a palette as a vector
jColors(palette = "basic", colors = c("red", "blue"))
}
\keyword{colors}
