% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jTheme.R
\name{jTheme}
\alias{jTheme}
\title{Theme settings for plotting using ggplot2}
\usage{
jTheme()
}
\description{
A custom theme for making plots in ggplot2.
}
\examples{
# Make example plot:
library(ggplot2)
data("midwest", package = "ggplot2")
plot <- ggplot(midwest, aes(x=area, y=poptotal)) +
    geom_point(aes(col=state, size=popdensity)) +
    geom_smooth(method="loess", se=F) +
    xlim(c(0, 0.1)) +
    ylim(c(0, 500000)) +
    labs(subtitle="Area Vs Population",
        y="Population",
        x="Area",
        title="Scatterplot",
        caption = "Source: midwest")
plot

# Add jTheme
plot + jTheme()

# Add jTheme but keep legend
plot + jTheme() + theme(legend.position="right")
}
