% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadMultiplePackages.R
\name{loadMultiplePackages}
\alias{loadMultiplePackages}
\title{Install and load multiple R packages.}
\usage{
loadMultiplePackages(pkgs)
}
\arguments{
\item{pkgs}{A character vector with the names of the packages to load.}
}
\description{
This function installs and loads multiple R packages.
It is the exact same function as the 'ipak' function by Steven Worthington,
found at https://gist.github.com/stevenworthington/3178163.
The function first checks to see if the list of packages are installed.
If they are not, it installs them and then loads them into the R session.
}
\examples{
loadMultiplePackages(c('Hmisc', 'ggplot2', 'dplyr', 'tidyr'))
}
