% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit_probs.R
\name{logit_probs}
\alias{logit_probs}
\title{Computes logit fraction given a set of coefficients and alternatives}
\usage{
logit_probs(coefs, newdata, obsID = NULL, ci = 0.95)
}
\arguments{
\item{coefs}{A matrix of coefficients or draws of coefficients.}

\item{newdata}{A data frame of a set of alternatives for which to compute
logit probabilities. Each row is an alternative.}

\item{obsID}{The name of the column that identifies each set of
alternatives. Defaults to \code{NULL}, in which case it assumes the newdata
are all one choice scenario.}

\item{ci}{The sensitivity of the computed confidence interval (CI).
Defaults to \code{ci = 0.95}, reflecting a 95\% CI.}
}
\description{
Returns a data frame of the predicted probabilities (with a confidence)
interval) for a set of coefficients (or coefficient draws) and a data frame
of alternatives.
}
