% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printConcatCode.R
\name{printConcatCode}
\alias{printConcatCode}
\title{Prints the code to concatenate a set of strings}
\usage{
printConcatCode(strings, lineBreaks = FALSE)
}
\arguments{
\item{strings}{A vector of strings.}

\item{lineBreaks}{Set to TRUE to print a new line in between each string}
}
\description{
This function prints the code to concatenate a set of strings.
}
\examples{
# Create a set of strings:
strings = c('one', 'two', 'one', 'three', 'two', 'four')

# Print code to create the concatenated set of all unique string values:
printConcatCode(unique(strings))

# Print code with new lines in between each string:
printConcatCode(unique(strings), lineBreaks=T)
}
