% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qrender.R
\name{qrender}
\alias{qrender}
\title{\code{quarto::quarto_render()}, but output file can move}
\usage{
qrender(input, output_file = NULL, ...)
}
\arguments{
\item{input}{A qmd file.}

\item{output_file}{Path to the output file.}

\item{...}{Other args passed to \code{quarto::quarto_render()}}
}
\description{
The default \code{quarto::quarto_render()} function can only render outputs
to the current working directory. This is a wrapper that allows it to
render to a different directory. It just copies it to the new location
after rendering
}
