% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_speed.R
\name{quantile_speed}
\alias{quantile_speed}
\title{Faster quantile()}
\usage{
quantile_speed(x, probs = c(0.1, 0.9), na.rm = FALSE)
}
\arguments{
\item{x}{numeric vector whose sample quantiles are wanted, or an
object of a class for which a method has been defined (see
also \code{'details'}). \code{NA} and \code{NaN} values are not allowed in
numeric vectors unless \code{na.rm = TRUE}.}

\item{probs}{numeric vector of probabilities with values in \verb{[0,1]}.
(Values up to \code{'2e-14'} outside that range are accepted and
moved to the nearby endpoint.)}

\item{na.rm}{logical; if \code{TRUE}, any \code{NA} and \code{NaN}'s are removed from \code{x}
before the quantiles are computed.}
}
\description{
A faster version of quantile(), copied from this gist:
https://gist.github.com/sikli/f1775feb9736073cefee97ec81f6b193
}
\examples{
quantile(x <- rnorm(1001)) # Extremes & Quartiles by default
quantile(x,  probs = c(0.1, 0.5, 1, 2, 5, 10, 50, NA)/100)
}
