% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto_render_move.R
\name{quarto_render_move}
\alias{quarto_render_move}
\title{\code{quarto::quarto_render()}, but output file is moved to \code{output_dir}}
\usage{
quarto_render_move(input, output_file = NULL, output_dir = NULL, ...)
}
\arguments{
\item{input}{Path to the input qmd file.}

\item{output_file}{The name of the output file. If using \code{NULL} then the
output filename will be based on filename for the input file.}

\item{output_dir}{Path to the output directory.}

\item{...}{Other args passed to \code{quarto::quarto_render()}}
}
\description{
The default \code{quarto::quarto_render()} function can only render outputs
to the current working directory. This is a wrapper that moves the rendered
output to \code{output_dir}.
}
