% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_zip.R
\name{read_zip}
\alias{read_zip}
\title{Download a zipped file from the web and return the temporary paths to each
file}
\usage{
read_zip(url)
}
\arguments{
\item{url}{A url to a zipped file}
}
\description{
This function downloads and locally stores a zipped file from a url and then
returns a list of the paths to each file
}
\examples{
# Download and view fuel economy data for all U.S. vehicles
# from 1980 to present (https://www.fueleconomy.gov/feg/download.shtml):
url <- 'https://www.fueleconomy.gov/feg/epadata/vehicles.csv.zip'
filePaths = read.zip(url)
print(filePaths)

# View first data file:
data <- read.csv(filePaths[1], header = T)
head(data)
}
