% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_df.R
\name{rep_df}
\alias{rep_df}
\title{Repeats a data frame n times}
\usage{
rep_df(df, n)
}
\arguments{
\item{df}{A data frame.}

\item{n}{The number of times to repeat the data frame by row.}
}
\description{
Returns a data frame repeated by row n time.
}
\examples{
df <- data.frame(
  one = c(1, 2),
  two = c(3, 4),
  three = c(5, 6)
)

rep_df(df, 3)

}
