% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortByCol.R
\name{sortByCol}
\alias{sortByCol}
\title{Sorts data frame or matrix by a desired column number.}
\usage{
sortByCol(df, col)
}
\arguments{
\item{df}{A data frame or matrix to sort.}

\item{col}{The desired column by which to sort.}
}
\description{
This function sorts a data frame or matrix by a desired column number.
It essentially does the same function as \code{arrange} in the plyr
package.
}
\examples{
sortByCol(cars, 2)
}
