% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_dir.R
\name{source_dir}
\alias{source_dir}
\title{Source all R files in a given directory.}
\usage{
source_dir(path, trace = TRUE, ...)
}
\arguments{
\item{path}{Path to directory.}

\item{trace}{Defaults to TRUE.}
}
\description{
This function sources all R files in a given directory.
}
\examples{
source_dir('filePath')
}
