% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourceDirFiles.R
\name{sourceDirFiles}
\alias{sourceDirFiles}
\title{Source all R files in a given directory.}
\usage{
sourceDirFiles(path, trace = TRUE, ...)
}
\arguments{
\item{path}{Path to directory.}

\item{trace}{Defaults to TRUE.}
}
\description{
This function sources all R files in a given directory.
}
\examples{
sourceDirFiles('filePath')
}
