% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_code.R
\name{vector_code}
\alias{vector_code}
\title{Prints the code to create a given vector}
\usage{
vector_code(a, line_breaks = FALSE)
}
\arguments{
\item{a}{A vector of elements for which the printed code will create.}

\item{line_breaks}{Set to TRUE to print a new line in between each string}
}
\description{
This function prints the code to create a given vector.
}
\examples{
# Create a set of strings:
strings <- c('one', 'two', 'one', 'three', 'two', 'four')

# Print code to create the set of all unique string values:
vector_code(unique(strings))

# Print code with new lines in between each string:
vector_code(unique(strings), line_breaks = TRUE)
}
