% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_html.R
\name{write_html}
\alias{write_html}
\title{Write a html file to disc}
\usage{
write_html(url, filename)
}
\arguments{
\item{url}{url to webpage.}

\item{filename}{Path to save html to disc.}
}
\description{
Save the html from a url to disc
}
\examples{
url <- "https://www.formula1.com/en/results.html/2022/drivers.html"
# Uncomment to save to disc:
# write_html(url, "formula1.html")

}
