% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cs.R
\name{Cs}
\alias{Cs}
\title{Vector of character strings from list of unquoted names}
\usage{
Cs(...)
}
\arguments{
\item{path}{Path to directory.}

\item{trace}{Defaults to TRUE.}
}
\description{
Returns a vector of character strings from list of unquoted names.
(Copied directly from the Hmisc package).
}
\examples{
Cs(a,cat,dog)
}
