% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installMyPackageLibrary.R
\name{installMyPackageLibrary}
\alias{installMyPackageLibrary}
\title{Install over 300 frequently-used packages.}
\usage{
installMyPackageLibrary(packageList)
}
\arguments{
\item{packageList}{An Rda file listing all the packages to install.}
}
\description{
This function installs a long list of packages that I frequently use.
I only use this after first installing R on a machine to get it ready for
use.
}
\examples{
load(file="./packageList.Rda")
installMyPackageLibrary(packageList.Rda)
}
