% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jColors.R
\name{jColors}
\alias{jColors}
\title{jColors}
\usage{
jColors(viewColors = FALSE)
}
\description{
Returns a list of different color palettes. Each palette is a data.frame.
trove Palette found at http://colrd.com/palette/19308/
redToGray Palette found at http://colrd.com/palette/21894/
}
\examples{
# View barplots of each color palette
colors = jColors(viewColors=T)
}
\keyword{colors}
